# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API Documentation",
        "description": "Dokumentasi API untuk project Laravel kamu",
        "version": "1.0.0"
    },
    "paths": {
        "/api/berita": {
            "get": {
                "tags": [
                    "Berita"
                ],
                "summary": "List semua berita",
                "operationId": "f32efb7f412a061f51fb3cca155c3c81",
                "responses": {
                    "200": {
                        "description": "Daftar berita",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Berita"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Berita"
                ],
                "summary": "Tambah berita baru",
                "operationId": "477776871d0c6397546a94e9fda672b6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "judul",
                                    "slug",
                                    "content",
                                    "kategori_id",
                                    "user_id",
                                    "status"
                                ],
                                "properties": {
                                    "judul": {
                                        "type": "string"
                                    },
                                    "slug": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "thumbnail": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "kategori_id": {
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "draft",
                                            "published",
                                            "archived"
                                        ]
                                    },
                                    "published_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berita berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Berita"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/api/berita/{slug}": {
            "get": {
                "tags": [
                    "Berita"
                ],
                "summary": "Detail berita",
                "operationId": "f38ca97a6123e40b0d75d5c85c6d3e75",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data berita",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Berita"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/berita/{id}": {
            "put": {
                "tags": [
                    "Berita"
                ],
                "summary": "Update berita",
                "operationId": "95981a76278fa3c11b1e282aa0526951",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "judul": {
                                        "type": "string"
                                    },
                                    "slug": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "thumbnail": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "kategori_id": {
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "draft",
                                            "published",
                                            "archived"
                                        ]
                                    },
                                    "published_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berita berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Berita"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Berita"
                ],
                "summary": "Hapus berita",
                "operationId": "aa77d202355e0df3e83a7dfe02111b4a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Berita dihapus"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Berita": {
                "title": "Berita",
                "description": "Model data berita",
                "properties": {
                    "id": {
                        "description": "ID berita",
                        "type": "integer"
                    },
                    "judul": {
                        "description": "Judul berita",
                        "type": "string"
                    },
                    "slug": {
                        "description": "Slug berita",
                        "type": "string"
                    },
                    "content": {
                        "description": "Isi berita",
                        "type": "string"
                    },
                    "thumbnail": {
                        "description": "URL thumbnail",
                        "type": "string",
                        "nullable": true
                    },
                    "kategori_id": {
                        "description": "ID kategori",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "ID user pembuat",
                        "type": "integer"
                    },
                    "status": {
                        "description": "Status berita",
                        "type": "string",
                        "enum": [
                            "draft",
                            "published",
                            "archived"
                        ]
                    },
                    "published_at": {
                        "description": "Tanggal publikasi",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Tanggal dibuat",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Tanggal diupdate",
                        "type": "string",
                        "format": "date-time"
                    },
                    "kategori": {
                        "$ref": "#/components/schemas/Kategori"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "Kategori": {
                "title": "Kategori",
                "description": "Model data kategori",
                "properties": {
                    "id": {
                        "description": "ID kategori",
                        "type": "integer"
                    },
                    "nama": {
                        "description": "Nama kategori",
                        "type": "string"
                    },
                    "slug": {
                        "description": "Slug kategori",
                        "type": "string"
                    },
                    "deskripsi": {
                        "description": "Deskripsi kategori",
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Tanggal dibuat",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Tanggal diupdate",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "Model data user",
                "properties": {
                    "id": {
                        "description": "ID user",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nama user",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email user",
                        "type": "string",
                        "format": "email"
                    },
                    "email_verified_at": {
                        "description": "Tanggal verifikasi email",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Tanggal dibuat",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Tanggal diupdate",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}